refactor: todo add, remove 함수 분리
